requirements.txt
# 截图库
mss==10.0.0
PyAutoGUI==0.9.54

# 图像处理
opencv-python==4.10.0.84
Pillow==10.4.0
numpy==2.1.3

# OCR 引擎
paddlepaddle==2.6.2
paddleocr==2.9.1

# 数据库
SQLAlchemy==2.0.35

# 数据分析
pandas==2.2.3

# 配置管理
PyYAML==6.0.2

# 飞书 API (可选)
requests==2.32.3

# GUI (可选)
PyQt6==6.7.1
PyQt6-Qt6==6.7.1

# 测试
pytest==8.3.3
pytest-cov==6.0.0
